refactored isLoggedIn and checkCampgroundOwnership middlewares
