+++ middleware/index.js
@@ -1,0 +1,31 @@
+const Campground = require('../models/campground');
+
+function isLoggedIn(req,res,next){
+    if(req.isAuthenticated()){
+        return next();
+    }else{
+        res.redirect("/login");
+    }
+}
+
+function checkCampgroundOwnership(req,res,next){
+    if(req.isAuthenticated()){
+        Campground.findById(req.params.id,function(err,foundCampground){
+            if(err){
+                res.redirect("back");
+            }else{
+                if(foundCampground.author.id.equals(req.user._id)){     // using .equals because first variable is an
+                                                                        // object while other is string so == or ===
+                                                                        // will not work
+                    next();
+                }else{
+                    res.redirect("back");
+                }
+            }
+        })
+    }else{
+        res.redirect("back");
+    }
+}
+
+module.exports = {isLoggedIn, checkCampgroundOwnership};
